requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
filelock==3.25.2
Flask==3.1.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
python-discovery==1.1.3
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==21.2.0
Werkzeug==3.1.6
WTForms==3.2.1
